requirements.txt
aiogram==2.21
aiogram==2.25.1
asyncpg==0.27.0
gino==1.0.1
pyngrok==5.1.0
pyngrok==6.0.0
python-dotenv==1.0.0
SQLAlchemy==1.3.24
